+++ libsaas_gitlab/merge_requests.py
@@ -2,7 +2,6 @@ from libsaas import http, parsers
 from libsaas.services import base
 
 from . import resource
-from . import notes
 
 class MergeRequestsBase(resource.GitlabResource):
     path = 'merge_requests'
@@ -35,19 +34,19 @@ class MergeRequest(MergeRequests):
 
         return http.Request('PUT', url, data), parsers.parse_json
 
-    @base.resource(notes.Note)
+    @base.resource(resource.Note)
     def note(self, note_id):
         """
         Return a resource corresponding to a single note.
         """
-        return notes.Note(self, note_id)
+        return resource.Note(self, note_id)
 
-    @base.resource(notes.Notes)
+    @base.resource(resource.Notes)
     def notes(self):
         """
         Return a resource corresponding to all notes.
         """
-        return notes.Notes(self)
+        return resource.Notes(self)
 
     @base.resource(MergeRequestComments)
     def comments(self):