+++ Changelog.md
@@ -1,3 +1,15 @@
+## v0.3
+
+* 1c2cd47 add system hook support
+* c02f3d7 add session support
+* 135fef5 add services support
+* 7e1377a add repository files support
+* 938e79b add repository support
+* 3900fb3 add snippets and extend notes support
+* 1f74b3c add milestone support
+* 08ffc80 add labels support
+* 88609d7 add groups
+
 ## v0.2
 
 * separate keys resource